backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
channels==4.2.0
channels-redis==4.2.1
redis==5.2.1
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
python-dotenv==1.0.0
drf-yasg==1.21.9
